cast last error to int before throwing

+++ cppcryptfs/cryptcontext.cpp
@@ -42,7 +42,7 @@ CryptContext::CryptContext()
 	m_mountEvent = CreateEvent(NULL, FALSE, FALSE, NULL);
 
 	if (!m_mountEvent)
-		throw(GetLastError());
+		throw((int)GetLastError());
 
 	m_config = new CryptConfig;
 }
